+++ fastcrypto-pq/Cargo.toml
@@ -16,6 +16,8 @@ fastcrypto-derive = { path = "../fastcrypto-derive", version = "0.1.4" }
 # The C backend (mldsa-native, pinned as a submodule inside the wrapper)
 mysten-mldsa-native-rs = { git = "https://github.com/MystenLabs/mysten-mldsa-native-rs.git", rev = "ea1a9794c7ad0c68b9410121964b7439540826bc", version = "0.1.0" }
 rand.workspace = true
+sha2 = "0.10.6"
+digest.workspace = true
 serde.workspace = true
 bincode.workspace = true
 
@@ -31,4 +33,8 @@ serde_json = "1.0.93"
 
 [[bench]]
 name = "mldsa65"
+harness = false
+
+[[bench]]
+name = "winternitz_ots"
 harness = false
