+++ fastcrypto/Cargo.toml
@@ -114,9 +114,6 @@ harness = false
 name = "bulletproofs"
 harness = false
 
-[[bench]]
-name = "winternitz_ots"
-harness = false
 required-features = ["experimental"]
 
 [features]
